Update patch set 1
No score.

The renaming of Sim to Icc seems somewhat gratuitous and it
might be better to have this as a separate CL if it was
deemed desirable. This would also reduce the size of this
CL somewhat which would be helpful.

As has been suggested overloading the Call and Connection
class names doesn't feel right and appending to the front
GSM/CDMA would be preferable to me.

Most of my other comments are simple formatting problems
that should be easy to resolve. Of course others should
be fixed as found.

Patch-set: 1