Update patch set 1
No score.

I've mainly been reviewing the changes under apps/Phone (change 4412),
but wanted to make one general comment here (which a couple of other
reviewers have also alluded to.)

Please don't use new class names ("CallBase" and "ConnectionBase") for
the base classes that used to just be called "Call" and "Connection".

(I realize some people do use that naming convention, but it's not
Android style.)

Instead, it would be much better for the current "Call" and "Connection"
classes (and "GSMCall" too, I guess) to just stay the way they are.  A
new CDMA Call class would then be "CDMACall".  That'll be more
consistent with the rest of the Android code base, *and* has the
additional benefit of touching far less code in the app layer (i.e.
change 4412...)

Patch-set: 1