New config parameter for provisioning.
This introduces a new (run-time) configuration parameter that specifies whether
the device requires provisioning, so that applications can adapt their behavior
based on that without having to know how that parameter is specified.

The current implementation is a simply wrapper around SystemProperties, for
simplification, but that's literally an implementation detail and can be changed
to any other scheme in the future.