Hacks to make the system start up without OpenCORE.
This shouldn't be submitted.

The basic idea is to remove the native libraries that link against OpenCORE,
except for libsystem_server.so where the call into such a library is removed.

Further hacks include not waiting for audioflinger to start, and removing media
classes from the list of preloaded classes so that zygote doesn't crash trying
to initialize them.