Adhoc Support for Wifi Connections
AP_SCAN mode will not be automatically set to 1 anymore.  AP_SCAN mode will be set based on the availability of adhoc signals in wpa_supplicant.conf

Adhoc signals will now be watched for in WifiService.java similar to the way Hidden networks are watched.

WifiConfiguration.java will now be able to set frequency and adhoc mode for a given configuration.