Initialize mVisibilityChanged to fix a bug
This is a fix for the camcorder application.

When we attempt to record the video for the second time,
we found that the overlay was not being setup correctly.
The function LayerBuffer::OverlaySource::onVisibilityResolved()
sets up the overlay. But it does not do so the second time,
because the variable mVisibilityResolved is not set to true.
We already discussed this issue with Mathias Agopian and he
agrees that it is an issue. This patch is a workaround until
a proper solution is implemented.

I also filed a bug:

http://code.google.com/p/android/issues/detail?id=2350

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>