Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Looks to be mostly modeled after the existing ogg player, so code wise it's probably OK. My only concern would be the license and size of the additional library, which are unknowns at this point (at least to me).

Also, at some point we were looking in to a using a different ogg library (since Tremor development is sporadic at best), which had the additional advantage of also supporting flac, so that might be an option too.

Patch-set: 1