Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

>I'm considering ShapeDrawable and ColorDrawable to be very similar in >their interfaces.

They actually are not. The goal of ColorDrawable is to fill the clip region of the Canvas, which is why it has a negative instrinsic width and height. ColorDrawable is very particular in that regard; even if you set the bounds of a ColorDrawable, it will always fill the entire clip region. All other drawables must have their bounds set correctly, and this include ShapeDrawable.

Patch-set: 1
Label: Code-Review=-1