Update patch set 1
You are right. If someone uses GradientDrawable and decides to use more advanced Shape, he should switch to ShapeDrawable and it that case he will need to set dimensions to -1x-1 to get the same behavior.

In any case I don't see much meaning leaving the field uninitialized and counting on the default. It's bad code style. In the worst case there should be a note in documentation, that ShapeDrawable is not useful when setting passing to a View as a Drawable in most cases, because of the default sizes and you should invoke setIntrinsicH/W() to change that.

Patch-set: 1