Update patch set 1
>In any case I don't see much meaning leaving the field uninitialized and >counting on the default. It's bad code style.

It's not bad code style in Java, it's understanding the language you are using :) Anyway, this is not the place to debate this, especially since this is something that is acceptable (and commonly used) on the Android project.

> In the worst case there should be a note in documentation, that >ShapeDrawable is not useful when setting passing to a View as a Drawable >in most cases, because of the default sizes and you should invoke >setIntrinsicH/W() to change that.

I agree, the documentation should reflect that. However, the proper way to do it is not necessarily to call setIntrinsicWidth/Height but to simply use setBounds(), similarly to other drawables.

Patch-set: 1