Update patch set 1
> I agree, the documentation should reflect that. However, the proper way to do it is not necessarily to call setIntrinsicWidth/Height but to simply use setBounds(), similarly to other drawables.

setBounds() is not the answer. If you pass ShapeDrawable to Button, ProgressBar or ImageView. They will set the bounds based on intrinsic dimensions, meaning to 0x0 by default.

You are right in case that you use ShapeDrawable.draw() directly. But that's not always the case. And when setting up bounds by yourself, you may need to use intrinsic dimensions too, like the mentioned Views do.

Patch-set: 1