Update patch set 1
Patch Set 1: No score

Patches to add WavPack support as per the following enhancement request:

http://code.google.com/p/android/issues/detail?id=4854
 
The code allows playback of WavPack files and includes support for reading the metadata (tag information) from the files including embedded album art. 

This covers 3 patches, 13007, 13008 and 13009

The source code for libWavPack is hosted on github

http://github.com/soiaf/android_external_wavpack

You can enable this by adding this to your .repo/local_manifest.xml

<remote name="github" fetch="git://github.com/" />
<project path="external/wavpack" name="soiaf/android_external_wavpack" remote="github" />

WavPack is BSD licensed. The generated (ARM) library file is based on the 4.60.1 codebase and is approximately 58K in size (there is scope to reduce this size by a few KB if required).

Patch-set: 1