Create change
Patch-set: 1
Change-id: Ica8657306d8c2ba37af86c8c1546d704033a088e
Subject: This patch fixed the order of digits in RTL text, Originally digits that were in RTL text were displayed in RTL order. (you can see that digits direction was set to the direction of the text before it) It is now always set to LRT, which is the way it should be.
Branch: refs/heads/master
Commit: ca8657306d8c2ba37af86c8c1546d704033a088e