Refactored the proposed GLWallpaperService and GLSurfaceView so that they share a common codebase yet remain completely backwards-compatible.
I put the common code for both into EglHelper and left it public so GLWallpaperService can use it in the wallpaper package and so that developers can use it to create GLSurfaces in other ways.
I added one additional public interface which is GLRenderer and is shared by GLSurfaceView and GLWallpaperService.  GLSurfaceView keeps its Renderer but that now extends GLRenderer so that any GLRenderer can be used in a GLSurfaceView and a GLWallpaperService interchangeably, which is something I found myself needing when I developed a Live Wallpaper.  It opens up the option for a game to include a live wallpaper component that uses the same renderer as the game itself.

I tested this code with my Live Wallpaper and with my latest game engine.  No problems occured.