The following changes improve the reporting done in HierarchyViewer via the ViewDebug interface.
Wind River has built a system testing platform for Android.  It is called STP - for System Testing Platform.
We use Android's adb to access ViewDebug in Android phones.  The ViewDebug interface is missing a few important
values that we can use to run tests.  These changes add access to those values, as described below;

view/ViewDebug.java:  This change adds a screenX and screenY position for the overall view.  Without these
values, we cannot tell where a popup window is on the screen.  We use Monkey's network interface to click the
Android screen in specific locations, and we need to know where windows are positioned.

widget/AnalogClock.java:  We'd like to know the time that the clock is showing to verify it.

widget/CheckedTextView.java:  We'd like to know if the item is checked on not.

widget/CompoundButton.java:  We'd like to know if the item is checked on not.

widget/ProgressBar.java:  we need to know isIndeterminate(), getProgress(), getSecondaryProgress(), and
getMax() so we can understand what sliders and progress bars are showing.

widget/TextView.java: we need to know the current selection:  getSelectionStart() and getSelectionEnd()

We've been testing versions of this code on Cupcake, Donut and Eclair on emulators, dev phones and other
devices.  We'd like these changes in soon to be able to run our suites without asking our partners to change
their Android sources and so we can test new phones without any patching, just by plugging them in.

Thank you for your consideration.