Avoid creating a SP from a dying SP, causing multiple deletion
Once an SP has hit mStrong = 0, it is in the process of being deleted, making
another SP from it will result in the object being deleted twice.

For some reason ProcessState uses a special function that can do just this and
causes bugs.   We reimplement it to do an attempt at creating and SP instead of
forcing, then it clean up appropriately.