Allow a surface to be re-parceled
When a surface is originally created in SurfaceFlinger, the system
returns a SurfaceControl object, which is a wrapper that can be
used to create a surface.  This gets parceled to the client process,
which uses the data to create an actual surface, but does not create
a SurfaceControl.  Thus, every process with access to a surface will
have either a Surface or a SurfaceControl, but never both.

However, the parceling code assumes that there is always a SurfaceControl,
and uses its information to perform the parcel.  This works for server-side
surfaces, which have a SurfaceControl, but if an attempt is made to re-parcel
a client-side surface, the SurfaceControl object is null, so the parcel fails.
However, a client-side object will contain a Surface object instead, which
still contains all the information necessary to perform the parcel.

This patch adds the ability to parcel from a Surface, in addition to a
SurfaceControl, meaning that a Surface received by another process can be
correctly re-parceled to a third process.

Change-Id: I18884e91232d033d9db607a594bf6bfee4224a27