Update patch set 4
Patch Set 4:

Wow that sounds harsh. Let me try and explain.

GLSurfaceView is a delicate class, and I don't want to take on the burden of analyzing your patch set to see if it breaks anything.

I also don't see the point of your GLWallpaperService having its own implementation of GLSurfaceView.

Please try to come up with a design that allows you to use the existing GLSurfaceView as-is, without copying it or refactoring it. GLSurfaceView is designed to be used and customized by composition rather than subclassing, so subclassing and refactoring should not be nescessary.

Patch-set: 4