Add support for incremental screen grabbing
This builds on a previous patch to provide support for grabbing only
the areas of the screen that have changed since the previous grab.
This greatly improves performance for applications such as VNC servers
that want to react to display changes.

A new boolean parameter 'incremental' has been added to grabScreen().
If true, the function blocks until the screen changes, rather than
returning immediately.

Another parameter 'updatedRegion' returns the changed area of the
screen to the calling application.