Update patch set 2
Patch Set 2:

Someone pointed out that the new file, part of frameworks/base and obviously not a fully independent module, comes with a Code Aurora copyright and a license that is not Apache 2.0, which is something that we want to avoid as it creates headaches for license compliance and for future contributions.

Naveen, could you please upload a change that replaces that copyright/license with the standard AOSP/ASL2 one? Otherwise I'll find myself forced to revert this change, which is something that I'd like to avoid.

(I realize that there are a few places that don't currently follow that rule. Those are tolerated, but not grandfathered in, and I want to take care of them in the future).

Patch-set: 2