Update patch set 2
Patch Set 2: No score; Do not submit

Surface should not be re-parcelable. It is not allowed to "resend" a surface to another process; in particular, a Surface is associated to a particular client, sending it to another process makes no sense unless the other process accesses the same client, which we definitely don't want to allow (the window-manager is a special case).

Patch-set: 2
Reviewer: Gerrit User 1001056 <1001056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0