Update patch set 2
Patch Set 2: No score

Perhaps this needs to turn into a larger discussion.  Part of our software architecture requires a (native) daemon process to be able to render into a SurfaceView owned by an application.  Everything works beautifully with this change in place, so the hope was that we would be able to do it this way, and get this change upstream.  If it's the case that we don't ever want surfaces to be able to escape their client process, is there some other mechanism by which this goal might be accomplished?  I understand it's an edge case, but it's one which we have found extremely useful.

Patch-set: 2
Reviewer: Gerrit User 1003080 <1003080@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0