Platform part of Language switch delivery
Due to political conflicts, we're not allowed to sell phones
with access to certain languages in certain countries.

For example, Macedonian may not be used in Greece.

This can be solved in Android by building software variants where
certain language resources are excluded.

However, this is solution to be able to avoid many varinats build.

It filters out locales based on excluded language codes
specified in a system property, "ro.product.language.excluded".

The languages should be set in a space delimited string, like
"de es nl" for german, spanish and dutch.

The filter is applied both when an application fetches a list of
supported system locales via AssetManager, and when an application
(like Settings, MoreLocales2) tries to set the system locale via
the hidden interface IActivityManager.

The getLocales() patch is simply that we filter out certain
locales before they are put in the Java vector. In the set locale
case, we null the configuration locale variable which basically
makes the operation a no-op for locale updates only. Note that
this is the only possible solution. There is no way to propagate
an exception or error code without vast patching in different places
of the platform.

Bug: 28054580

Change-Id: I41987a6b8c55a737e0c48ad5a23c2a4e86a861a0