Update patch set 1
Patch Set 1:

I agree with you, that sometimes a custom dialog is better than a toast.
I personal use for this would be this:
I have an app, that requires an sdcard (without one, it wont run cause it needs it to download content and a db get stored on it). When the User now mounts the sdcard to the computer or removes it, the app will crash (on db.open cause sdcard not mounted). So i implemented a check, and if theres no State.MOUNTED, i will display a custom toast to the user and the app exits.
Im not a friend of a dialog in this case, cause theres no need for the app to run to inform the user, and since there can be custom layout toasts (via layoutinflater), it would be great to display it for a longer time.
I think this value should not be controllable by the dev so a app can not display a toast for 5 minutes or so, but i think 5 seconds is a good maximum value to display a bigger toast.

Patch-set: 1