Update patch set 1
Patch Set 1:

Well, I'm not the one to make the call on this, but I doubt something like this would be approved.

In my experience, success in Open Source is often about saying "no" as much as it is about "yes", otherwise we'll soon have Toast.LENGTH_EVEN_LONGER and Toast.A_BIT_LONGER_THAN_LONG_BUT_LESS_THAN_EVEN_LONGER.

It's a slippery slope.

Patch-set: 1