Update patch set 1
Patch Set 1: No score; Do not submit

I'm the one who gets to make the call, and I actually agree with Brad, for the reasons he gave.  The UX that I'd suggest is to have an activity that says something along the lines of "Insert an SD card to proceed," or something like that.  Just quitting with a toast, no matter how long it stays up isn't really right.  The toast will be too long for some and not long enough for others.

-joe

Patch-set: 1
Reviewer: Gerrit User 1000701 <1000701@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0