Update patch set 4
Patch Set 4:

Whoops, I think that's my fault.

The first version of the patch was _tiny_ but I asked that an expression be hoisted out into a renamed variable for clarity, but Johan then went all the way and renamed all the variables with explicit "seconds" and "milliseconds" suffixes for extra-clarity, but perhaps that went too far.

Should we test that this can merge as-is, or go back to a simpler version of the patch?

Patch-set: 4