Update patch set 2
Patch Set 2:

A master on/off could be a good idea since enabling cell broadcast consumes battery, and perhaps the user wants to have control over that. Such a setting would however prevent individual apps listening to CB from functioning properly, but it does feel quite reasonable to have it anyway (just as turning on airplane mode in the central settings would affect apps using connectivity).

Furthermore, I'm thinking that if a central CB menu is desired, that would probably relate to "normal" CB (i.e. the ones that are supposed to pop up or be shown as BSID), right? In that case, such a menu could be realized with the current API suggestion. Would it necessarily be interesting to show _all_ channels currently being listened to by _any_ app in such a menu? If a certain app uses CB as bearer, it's likely that it's for proprietary data messages of no interest to anyone else, and in that case shouldn't the subscription to that channel be dealt with by that app alone?

But sure, if there is need for a central override mechanism for enabling/disabling channels (or just listing all the enabled ones), that could be added. Using a special permission does sound like a good idea in that case.
Or perhaps such an addition could be made at a later stage, if really needed?

Patch-set: 2