Update patch set 4
Patch Set 4:

Jeff and Jeff, Patch Set 3 had been approved, and Patch Set 4 is essentially the same as 3. There had been some discussion around a potential simplification, but it seems to me that a conclusion was reached that the simplification wouldn't work.

Should I go ahead, carry the approval forward, and submit this change?

Patch-set: 4