Update patch set 1
Patch Set 1: No score; Do not submit

The impact is that any ImageButton created programmatically is now non-focusable by default. That's pretty severe. The "focusable" property is, to us, an intrinsic property of a widget. Ideally it should not be overridden. It's also why it's set from code, for the case when widgets are created programmatically. It's made difficult to override on purpose (to do so, you'll have to call setFocusable(false) from code, after the widget is created.)

Patch-set: 1
Reviewer: Gerrit User 1001146 <1001146@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0