Added method getPreferredHttpHost
Added method for returning preferred proxy which takes both
localhost and WIFI into account. This is a convenient method
to clients which only wants to set a correct proxy and don't
want to build in dependency to if WIFI is active or not.
Currently no WIFI proxy is supported by the system, but once
added, this method could return a suitable proxy for WIFI.

Change-Id: I8c9c2879351fd25a20ea82a2cb000f226248c357