Added a simple count() function to SQLiteDatabase
This method can be used to simply get the count of rows of a table.
You can also specify where arguments to filter the query.
This is more a helper function to quickly get a recordcount, without
opening a cursor and doing all the related stuff.
An example for this would be, to check if a table contains at min
1 row.
UnitTests for this 3 new methods are added to the cts project

Change-Id: Id9233aec0eaac08839041ae7cbaba203470ad3d8