Create change
Patch-set: 1
Change-id: I489cc453bdce7cb376102cba67a9ea470e13292d
Subject: system_server can potentially monitor uevent devices that are of no use to it. For instance, an accelerometer implementation as uevents. This would cause the process to be busy when unnecessary. If a device cannot be classified, don't monitor it.
Branch: refs/heads/master
Commit: e911044a23a8818e97dd03e13143eedd11e4eaa7