Update patch set 2
Patch Set 2: Do not submit

(a) The default settings should be configured as desired, and the setup wizard should be using StatusBarManager.disable() to turn off alarms.  If this is not already the case, we should fix it.  I've never heard of this bug before, but if it is indeed happening, there's a problem somewhere else that's causing it.

(b) I don't want to use DEVICE_PROVISIONED anywhere else in the system.  It was a mistake to add it in the first place.

Patch-set: 2
Reviewer: Gerrit User 1000701 <1000701@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0