Update patch set 1
Patch Set 1: Fails

(6 inline comments)

Thanks!

There are a few style issues, which I mentioned in one of the files but also apply in other places. We tend to be picky about those (though we prefer to not fix such issues once they've been submitted, in order to avoid conflicts, so it's a bit "do as we say, not as we do").

In good news, the code merges and builds fine in the open-source tree, and merges fine into Google's internal master tree which is a bit further ahead.

Patch-set: 1
Label: Code-Review=0
Label: Verified=-1