Update patch set 1
Patch Set 1:

When a service’s process dies, InputMethodManagerService sends MSG_UNBIND to InputMethodManager where it clears the binding. 
But, in the case where mBindSequence is not equal to mCurSeq, it fails to clear the binding and this results to continuous DeadObject Exceptions on the subsequent keypresses.

startInputLocked() in InputMethodManagerService.java increments mCurSeq for every client even if the InputBindResult is null.


But, InputMethodManager assigns mCurSeq to mBindSequence only when InputBindResult is non-null. 

When onServiceDisconnected() is called, mCurSeq is passed in UNBIND message, and InputMethodManager clears the binding only when mBindSequence is equal to mCurSeq. 
So this will have a left out binding of the service’s process killed.

Scenario to reproduce: 
1. Go to Settings->Language & Keyboard -> Select Language
2. Kill com.android.inputmetod.latin from shell 
3. Now press “Back” Key
This resulted in DeadObject exception as it tried to dispatch the key event to the binder for which the remote process has already died.
The situation occurs in Low Memory Conditions when com.android.inputmetod.latin gets killed by Low Memory Killer.

We suggest the following solutions:
 1. To increment mCurSeq only when mCurrentMethod is null (needResult = true) in InputMethodManagerService.java.
 2. To move clearBindingLocked() call outside  “if (mBindSequence == sequence) {“  check in InputMethodManager.java.

Please suggest if there is a better way to do this

Patch-set: 1
Reviewer: Gerrit User 1004300 <1004300@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0