Stagefright: Handling playback negative seek scenario
This patch has change to fix the issue while performing the stagefright
command line "b 3" option tests. They were failing as negative offset was
calculated and used for seek. In the SF middleware (OMX Codec::read),
we have condition to check the parameters. If the seek value is < 0,
then android assert is triggerred which leads to the issue.
Actually as per the test app, the test has to seek and continue decoding
till the EOF. Hence have modified the seek position handling accordingly.

Change-Id: I0be94f876f3ab3671a609ef96e0b1b716f22e4aa
Signed-off-by: Prasanna KumarM.R. <prasanna.kumarmr@ti.com>