Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

the only thing special about Float.compare is its handling of NaNs and infinities, neither of which is relevant here. i can't reproduce your claim of incorrect behavior from "f >= 0" with f equal to -2368302.0 (as mentioned in the check-in comment) or -2368314.0 (as mentioned in the review comment).

do you have a stand-alone example showing the wrong behavior? (which would appear to be a VM bug we should look at, rather than something to work around here, unless i've misunderstood...)

Patch-set: 1