Update patch set 1
Patch Set 1:

I will try to provide more details with some more experiments. Meanwhile, i see a variable dir declared as a member variable in TrackballAxis and the same name is used for a local variable (final int dir) in the loop of TrackballAxis::generate. This seems to cause some confusion, is this on purpose?

Patch-set: 1