Update patch set 1
Patch Set 1:

These apis were added so that we could change the visibility of the wallpaper from the Launcher. When the Launcher is displayed on the screen, the wallpaper is also alive and consumes CPU. The Launcher is unable to all the onVisibilityChanged as the WallpaperManager does not support the same.

Patch-set: 1