Telephony: Update the Mcc Table for India
The operator’s id (MNC) isn't displayed correctly because
it's cut-off to only 2 digits. This is a problem when creating
new APN's because this information is not known to the most users.
That means that after the creation of a new APN, the MNC code
will be wrong (only the first 2 digits are retrieved from the
table) and the user will not be able to access the internet.

Change-Id: Icc2aaa9717501d4d6cd2b9f829c342baa890695e