Update patch set 1
Patch Set 1:

Still -1.

Reread the change description.

We can't accept this change.  There are mnc codes for this mcc that only have 2 digits (en.wikipedia.org/wiki/Mobile_Network_Code) so the minimum number of mnc digits is 2.

This can be a problem for some carriers if they don't properly setup the SIM, but otherwise we couldn't access other carriers (those with 2 digits).

Patch-set: 1