Update patch set 1
Patch Set 1: Fails

Thanks for the contribution.

Before we can work any further, you're going to need to resolve a small issue with the copyright/license notices: every file that can have such a notice needs to have an apache 2.0 license header, with a single copyright line for The Android Open Source Project, with a single copyright year matching the year when the source file was originally written.

packages/WAPPushManager/tests/Android.mk is a good example.

Also, while you're at it, there are quite a few trailing whitespaces scattered along the source tree: they show up as bright red in the review pages so make sure that we don't miss them. Please take care of those as well.

Finally, because this is quite a complex contribution, it would probably make sense to also discuss it on the android-contrib list to better get an idea of the level of interest around adding wap push support in the core Android.

Patch-set: 1
Label: Code-Review=0
Label: Verified=-1