Update patch set 2
Patch Set 2: (2 inline comments)

What started this effort originally was a defect reporting that the add to dictionary dialog was not visible in asian languages. Removing "other" might be too limiting for some languages, questioning the entire patch, but it may also be a matter of taste. We'll investiagte internally, do you have some strong oppinions?

The second part is about automatically deselecting which seems like the right thing to do, and I assume there is still some place in the new code where this would be applicable?

Patch-set: 2