Telephony: Implement synchronous sending of sms messages
The code reads config_smsSyncSending resource
to determine if synchronous sending algorithm shoud be used.
When it is set to true SmsDispacher will wait for a message to be
sent before sending request for subsequent message while queueing all
messages that are requested to be sent. When this resource is set
to false (default) the SmsDispatcher behaves as it used to by sending
all messages as soon as it receives the request to send them.

Change-Id: I1e80ff4eede936a95e1a96d54bfb66dd4008dcd1