Update patch set 3
Patch Set 3:

There is similar fix already in MmsApp in Froyo.
Here is email discussion with Wei:

Alex,

the work was added in Android 2.2. The motivation was that when sending a
SMS message to multiple destinations (broadcast), we'd flood the RIL driver.
So we implemented queuing. We'd send the SMS to 1 destination, and put the
rest on the queue. When we receive confirmation from Telephony manager that
the one is sent, we'd send another, and so forth.

You are welcome to look at the source in MMS app. The relevant files are

SmsMessageSender.java
- sendMessage: calls queueMessage()
- queueMessage: puts all the messages in the sms db, marked them as
"queued", send broadcast intent w/ action
""com.android.mms.transaction.SEND_MESSAGE" to tell SmsReceiverService to
send the first in the queue.

SmsReceiverService.java
- sendFirstQueuedMessage: grab the first message off the queue, and
call SmsSingleRecipientSender.sendMessage()

SmsSingleRecipientSender.java
- sendMessage: calls smsManager.sendMultipartTextMessage() to actually send
the message

This code was mostly intended for sending to multiple recipients. However,
it should work to throttle repeatly sending messages to single recipient as
well.

Wei


On Wed, Sep 29, 2010 at 5:30 PM, <ayakav@codeaurora.org> wrote:

> Hi Wei,
>
> I wasn't aware of this new MMS app feature. I guess we can just abandon my
> change then.
>
> Thanks,
> Alex
>
> > first of all, MMS app now queue up outgoing messages to be sent, and send
> > them one at a time. When MMS app gets the sent notification from
> > TelephonyManager, it will send the next one in the queue.
> >
> > I guess that logic is in MMS app instead of Telephony Manager, so other
> > SMS/MMS apps in the Market couldn't take advantage of that.
> >
> > For the change in SmsDispatcher, John should take a look and see if it's
> > the
> > right place to do it.
> >
> > Wei

Patch-set: 3