Allow data connections after a no-reattach message from the network
In the current code there is a check whenever an application tries
to setup a data connection if the connection is automatically
attached by the platform or not. However, this check doesn’t take
into account that the platform might be disconnected by the network
together with a return code which asks it not to attach again.

This fix solves the problem when the network sends a de-attach
request with the command that re-attach is not required. This
would cause the network to never re-attach due to the noAutoAttach
property. Gprs state check was removed as well. When trySetupData
is called, the phone will try to attach if possible.

Change-Id: Ia09b80bc7f724c412030dbb3d40802c0b07f4f42