StagefrightMediaScanner: Call endFile() for MIDI and OGG files.
addStringTag() caches non-ascii metadata strings for later
processing, and then endFile() will be called at the end of
processFile() to convert non-ascii strings from locale's charset
to utf-8 if required.

Stagefright's processFile() failed to call endFile() when the
processing file is a MIDI file or an OGG file.  This patch fixes
this problem to populate metadata correctly.

Reviewed by: Brad Fitzpatrick, Marco Nelissen.
Thank you for your reviews.

Change-Id: I072e79d81dce1fec63297d2b5d2b870a72e5b66e