Force Android to use the offset in NITZ
If Android has no country-code set, and receives a NITZ-message,
it will wait with adjusting the timezone-offset until it gets a
proper country-code. This fix will make Android, when it has no country-code
set, use the offset in the NITZ message to temporarily guess a country and timezone.
The timezone and country will update when the network sends a country-code.

Change-Id: I2d7c1512d95ac74ea3c09ab18f9ac9d1e83fa1d9

Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>